requirements.txt
# Web Framework
fastapi
uvicorn

# Database & ORM
sqlalchemy

# AI & Agent Orchestration
langchain
langchain-openai
langchain-community
langgraph

# Utilities
pydantic
python-dotenv